+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **app**: added wire frame folder to hold penpot design wire frames
 - **app**: added default image gimp file
 - **app**: added error image for images that do not load properly
+- **app**: added filter.ts file to hold all the thread filters
 
 ### Fixed